refactor(view) : View를 inputView, resultView 으로 분리함.
