+++ src/main/java/lotto/LottoApplication.java
@@ -1,19 +1,20 @@
 package lotto;
 
 import lotto.domain.LottoResult;
+import lotto.domain.LottoTicket;
 import lotto.domain.LottoTickets;
-import lotto.domain.LottoWinningTicket;
+import lotto.view.LottoMachineResultView;
 import lotto.view.LottoMachineView;
 
 public class LottoApplication {
     public static void main(String[] args) {
         LottoTickets lottoTickets = LottoMachineView.lottoPurchaseView();
-        LottoMachineView.showTicketNumber(lottoTickets);
+        LottoMachineResultView.showTicketNumber(lottoTickets);
 
-        LottoWinningTicket lottoWinningTicket = LottoMachineView.latestWinningLotto();
-        LottoResult lottoResult = LottoMachineView.checkWinningResult(lottoWinningTicket, lottoTickets);
+        LottoTicket winningTicket = LottoMachineView.latestWinningLotto();
+        LottoResult lottoResult = LottoMachineView.checkWinningResult(winningTicket, lottoTickets);
 
-        LottoMachineView.printLottoResult(lottoResult);
+        LottoMachineResultView.printLottoResult(lottoResult);
 
 
     }
