+++ src/main/java/lotto/view/LottoMachineView.java
@@ -1,8 +1,9 @@
 package lotto.view;
 
-import lotto.domain.*;
-import lotto.exception.MinimumPurchaseAmountException;
-import lotto.util.LottoTicketUtils;
+import lotto.domain.LottoMachine;
+import lotto.domain.LottoResult;
+import lotto.domain.LottoTicket;
+import lotto.domain.LottoTickets;
 
 import java.util.Arrays;
 import java.util.List;
@@ -26,15 +27,10 @@ public class LottoMachineView {
         return new LottoMachine().purchaseTicket(purchaseAmount);
     }
 
-    public static void showTicketNumber(LottoTickets lottoTickets) {
-        for (LottoTicket ticket : lottoTickets.ticketList()) {
-            LottoMachineView.show(ticket.showTicketNumber());
-        }
-    }
 
-    public static LottoWinningTicket latestWinningLotto() {
+    public static LottoTicket latestWinningLotto() {
         System.out.println(WINNING_LOTTO_INFO_MESSAGE);
-        return new LottoWinningTicket(
+        return new LottoTicket(
                 Arrays.stream(split(LOTTO_SPLIT_DELIMITER))
                         .map(number -> number.trim())
                         .mapToInt(Integer::parseInt)
@@ -43,24 +39,9 @@ public class LottoMachineView {
         );
     }
 
-    public static LottoResult checkWinningResult(LottoWinningTicket lottoWinningTicket, LottoTickets lottoTickets) {
+    public static LottoResult checkWinningResult(LottoTicket lottoWinningTicket, LottoTickets lottoTickets) {
         return LottoMachine.winningResult(lottoWinningTicket, lottoTickets);
     }
-
-    public static void printLottoResult(LottoResult lottoResult) {
-        System.out.println("당첨 통계");
-        System.out.println("---------");
-        for (int i = 1; i < LottoPrize.values().length; i++) {
-            LottoPrize prize = LottoPrize.values()[i];
-            System.out.print(String.format("%d개 일치", prize.getMatchCount()));
-            System.out.print(String.format(" (%d원)", prize.getPrice()));
-            System.out.println(String.format(" - %d개", lottoResult.prizeCount(prize)));
-        }
-        System.out.println(String.format("총 수익률은 %.2f", lottoResult.average()));
-    }
-
-
-
 
     private static String[] split(String delimiter) {
         return sc.nextLine().split(delimiter);
